AI: Swagger and Psych Up
Some minor tweaks for Swagger. However, the changes in Psych Up are
notable. Now the AI will not use Psych Up if the target has no stat
beyond the ones of the user, as it used to simply spam Psych Up against
an opponent. Also, the AI might use Psych Up when the opponent is at +2,
while it previously required at least +3 in something.