AI: finer criteria for switching out when intoxicated
Essentially, this considers switching out sooner, but not too soon, when
intoxicated, if there are good switchins available. The code is also a
bit cleaner and has been tuned a bit. The logic might still not be easy
to follow, though.
This was done a long time ago and has been extensively tested. There is
a nonzero probability of other uncommited changes in this file being
necessary so that this modified code works, but they are expected to
eventually be commited.